break out rules engine from audience specifics
