+++ src/audience.ts
@@ -2,32 +2,24 @@
 // See Audience.md in docs/
 //
 
-/**
- * Evaluating an audience can yield a runtime error.
- */
-export type AudienceError = {
-    message: string;
-};
+import {
+    AST,
+    Atom,
+    Environment,
+    evalAst,
+    isError,
+    parse,
+    PrimitiveFn,
+    RulesEngineError,
+} from "./rules-engine";
 
-/**
- * The rules syntax is a subset of JSON.
- */
-export type AST = Atom | List;
-export type Atom = string | number | boolean;
-export type List = Array<List | Atom>;
-
-/**
- * The environment is passed in from the outside when evaluating rules
- */
-export type Environment = {
-    attributes: Record<string, Atom>;
-};
+type AudienceError = RulesEngineError;
 
 /**
  * Audience contains rules to be evaluated for activating projects.
  */
 export class Audience {
-    ast: AST;
+    rules: AST;
 
     /**
      * Create a new Audience with the given rules.
@@ -36,14 +28,13 @@ export class Audience {
      * @param rules JSON string describing the audience rules, see docs/Audience.md
      */
     constructor(rules: string) {
-        let ast: AST;
-        try {
-            ast = JSON.parse(rules);
-        } catch (x) {
-            throw { message: "rules syntax error" };
+        const result = parse(rules, Object.keys(primitives));
+
+        if (isError(result)) {
+            throw result;
         }
-        checkSyntax(ast);
-        this.ast = ast;
+
+        this.rules = result;
     }
 
     /**
@@ -51,7 +42,7 @@ export class Audience {
      * the audience matches.
      */
     eval(env: Environment): boolean | AudienceError {
-        const result = this.evalAst(this.ast, env);
+        const result = evalAst(this.rules, env, primitives);
 
         if (isError(result)) {
             return result;
@@ -63,84 +54,6 @@ export class Audience {
 
         return result;
     }
-
-    private evalAst(ast: AST, env: Environment): Atom | AudienceError {
-        switch (typeof ast) {
-            case "number":
-                return ast;
-            case "string":
-                return ast;
-            case "boolean":
-                return ast;
-            case "object":
-                if (ast instanceof Array && typeof ast[0] == "string") {
-                    return this.evalApply(ast[0], ast.slice(1), env);
-                }
-        }
-        return { message: `cannot evaluate ${ast}` };
-    }
-
-    private evalApply(head: string, args: List, env: Environment): Atom | AudienceError {
-        // TODO(Fabian) we have some extensibility in mind for primitives,
-        // might want to rephrase this or refactor where the lookup is.
-        const prim = primitives[head];
-        if (!prim) {
-            return { message: `${head} is not a primitive` };
-        }
-        try {
-            const evaledArgs: Atom[] = [];
-            for (const arg of args) {
-                const a = this.evalAst(arg, env);
-                if (isError(a)) {
-                    return a;
-                }
-                evaledArgs.push(a);
-            }
-            return prim(evaledArgs, env);
-        } catch {
-            return { message: `error applying primitive ${head}` };
-        }
-    }
-}
-
-function checkSyntax(ast: AST): void {
-    switch (typeof ast) {
-        case "object":
-            if (ast instanceof Array) {
-                checkSyntaxInner(ast);
-            }
-            return;
-    }
-    throw { message: "AST root must be a list" };
-}
-
-function checkSyntaxInner(ast: AST): void {
-    switch (typeof ast) {
-        case "number":
-            return;
-        case "string":
-            return;
-        case "boolean":
-            return;
-        case "object":
-            if (ast instanceof Array) {
-                const car = ast[0];
-                if (typeof car != "string") {
-                    throw { message: `can only apply strings, ${car} is not a string` };
-                }
-                // TODO(Fabian) not sure yet if we want to test this statically,
-                // we have some extensibility in mind with regards to callable functions.
-                if (!primitives[car]) {
-                    throw { message: `${car} is not a primitive` };
-                }
-                const cdr = ast.slice(1);
-                for (const elem of cdr) {
-                    checkSyntaxInner(elem);
-                }
-                return;
-            }
-    }
-    throw { message: `rules syntax error at ${JSON.stringify(ast)}` };
 }
 
 function stringFun(a: Atom, b: Atom, op: (x: string, y: string) => Atom): AudienceError | Atom {
@@ -157,8 +70,7 @@ function numberFun(a: Atom, b: Atom, op: (x: number, y: number) => Atom): Audien
     return op(a, b);
 }
 
-// eslint-disable-next-line @typescript-eslint/ban-types
-const primitives: Record<string, Function> = {
+const primitives: Record<string, PrimitiveFn> = {
     // boolean operations
     not: (args: Atom[]) => {
         const [arg] = args;
@@ -240,8 +152,4 @@ function getInEnvBool(name: Atom, env: Environment): boolean | AudienceError {
         return { message: `'${name}' is not a boolean` };
     }
     return val;
-}
-
-function isError(x: Atom | AudienceError): x is AudienceError {
-    return (x as AudienceError).message !== undefined;
 }
